+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### General
 
+- add pytest for get_wf_files ([#3725](https://github.com/nf-core/tools/pull/3725))
 - Update GitHub Actions (major) ([#3849](https://github.com/nf-core/tools/pull/3849))
 - add optional link to blogpost to sync PR ([#3852](https://github.com/nf-core/tools/pull/3852))
 - update multiqc version to fix utils test ([#3853](https://github.com/nf-core/tools/pull/3853))